Convert to MONOREPO (#28)
* move backend to backend directory

* move dot files to backend

* cd to backend

* add conditionals to only run back-end build and deployment only if back-end changes

* change README.md title

* test

* remove space

* only detect local changes

* add test change

* whatever
