--- .dockerignore
@@ -1,11 +1,0 @@
-bin/
-!bin/*.keep
-
-README.md
-
-.env
-.env.example
-
-.git/
-.github/
-.gitignore
