--- .env.example
@@ -1,2 +1,0 @@
-export PORT=8080
-export SRV_DIR=/srv
