+++ .github/workflows/push.yaml
@@ -2,29 +2,33 @@ name: Branch Push
 on: push
 jobs:
   changes:
+    name: Detect Changes
     runs-on: ubuntu-latest
     permissions:
       pull-requests: read
     outputs:
-      changes: ${{ steps.filter.outputs.src }}
+      backend: ${{ steps.filter.outputs.backend }}
     steps:
     - uses: actions/checkout@v4
     - uses: dorny/paths-filter@v3
       id: filter
       with:
         filters: |
-          src:
-            - 'go.mod'
-            - 'cmd/**'
-            - 'internal/**'
+          backend:
+            - 'backend/go.mod'
+            - 'backend/cmd/**'
+            - 'backend/internal/**'
+            - 'backend/.dockerignore'
+            - 'backend/Dockerfile'
+            - 'backend/scripts/**'
+            - 'backend/configs/**'
             - '.github/workflows/push.yaml'
-            - '.dockerignore'
-            - 'Dockerfile'
 
   convert-branch-name:
-    runs-on: ubuntu-latest
     needs: changes
-    if: ${{ needs.changes.outputs.changes == 'true' }}
+    name: Convert Branch Name
+    runs-on: ubuntu-latest
+    if: ${{ needs.changes.outputs.backend == 'true' }}
     outputs:
       name: ${{ steps.convert.outputs.name }}
     steps:
@@ -34,9 +38,11 @@ jobs:
         run: |
           output=$(echo "$BRANCH_NAME" | tr '[:upper:]' '[:lower:]' | sed 's/[^a-z0-9]/-/g' | sed 's/--*/-/g' | sed -E 's/^-+|-+$//g')
           echo "name=$output" >> $GITHUB_OUTPUT
-  build:
-    runs-on: ubuntu-latest
+  build-backend:
     needs: convert-branch-name
+    name: Build Backend
+    runs-on: ubuntu-latest
+    if : ${{ needs.changes.outputs.backend == 'true' }}
     env:
         BRANCH_NAME: ${{needs.convert-branch-name.outputs.name}}
     steps:
@@ -47,19 +53,19 @@ jobs:
         username: ${{ vars.DOCKERHUB_USERNAME }}
         password: ${{ secrets.DOCKERHUB_TOKEN }}
     - name: Build image
-      run: make docker
+      run: cd backend && make docker
     - name: Push image with commit sha
-      run: make docker-push TAG=$BRANCH_NAME-$(git rev-parse --short "$GITHUB_SHA")
+      run: cd backend && make docker-push TAG=$BRANCH_NAME-$(git rev-parse --short "$GITHUB_SHA")
     - name: Push image with latest tag
       run: >-
         if [[ $BRANCH_NAME == "main" ]]; then
-          make docker-push TAG=latest
+          cd backend && make docker-push TAG=latest
         else
-          make docker-push TAG=$BRANCH_NAME-latest
+          cd backend && make docker-push TAG=$BRANCH_NAME-latest
         fi
-  deploy:
-    needs: [build, convert-branch-name]
-    name: Create Deployment
+  deploy-backend:
+    needs: [build-backend, convert-branch-name]
+    name: Create Backend Deployment
     runs-on: ubuntu-latest
     env:
       BRANCH_NAME: ${{needs.convert-branch-name.outputs.name}}
