--- .gitignore
@@ -1,7 +1,0 @@
-**.env
-!.env.example
-
-.DS_Store
-
-bin/*
-!bin/.keep
