+++ backend/.env.example
@@ -1,0 +1,2 @@
+export PORT=8080
+export SRV_DIR=/srv
