+++ backend/.gitignore
@@ -1,0 +1,7 @@
+**.env
+!.env.example
+
+.DS_Store
+
+bin/*
+!bin/.keep
