+++ backend/go.mod
@@ -1,0 +1,3 @@
+module boshi-backend
+
+go 1.23.0
