--- configs/client-metadata-template.json
@@ -1,12 +1,0 @@
-{
-	"client_id": "https://<URL>.deguzman.cloud/client-metadata.json",
-	"client_name": "Boshi",
-	"client_uri": "https://<URL>.deguzman.cloud",
-	"redirect_uris": ["https://<URL>.deguzman.cloud/"],
-	"grant_types": ["authorization_code", "refresh_token"],
-	"scope": "atproto",
-	"response_types": ["code"],
-	"token_endpoint_auth_method": "none",
-	"application_type": "web",
-	"dpop_bound_access_tokens": true
-}
