--- go.mod
@@ -1,3 +1,0 @@
-module boshi-backend
-
-go 1.23.0
